+++ SampleApps/SPMTest/SPMTest/ViewController.swift
@@ -14,9 +14,10 @@ import BraintreeSEPADirectDebit
 class ViewController: UIViewController {
 
     override func viewDidLoad() {
+        // TODO: remove in the final PR for making authorization internal
         let apiClient = BTAPIClient(authorization: "sandbox_9dbg82cq_dcpspy2brwdjr3qn")!
 
-        let amexClient = BTAmericanExpressClient(apiClient: apiClient)
+        let amexClient = BTAmericanExpressClient(authorization: "sandbox_9dbg82cq_dcpspy2brwdjr3qn")
         let applePayClient = BTApplePayClient(apiClient: apiClient)
         let cardClient = BTCardClient(apiClient: apiClient)
         let dataCollector = BTDataCollector(apiClient: apiClient)