+++ CHANGELOG.md
@@ -29,6 +29,8 @@
     * Remove `fetchPaymentMethodNonces` methods and parser
   * BraintreePayPal
     * Update PayPal app URL query scheme from `paypal-app-switch-checkout` to `paypal`
+  * BraintreeAmericanExpress
+    * Update initializer to `BTAmericanExpressClient(authorization:)`
 
 ## 6.27.0 (2025-01-23)
 * BraintreePayPal